# Faster check if pojo has property change listener support.