Checking if class has `addPropertyChangeListener` via `Class#getMethod` is very slow, as it throws exception, populating its stacktrace is quite an expensie operation. In many usecases all objects injected to session may not have this method, so this operation is not rear. Looking for method by name is faster and increases overall performance. In our application this gave ~ 1.7 better throughput.