@mariofusco 2016-02-24
I'm not sure this will be a performance improvement in all cases. It really depends on how many methods the bean under inspection has. I suspect that if there are many of them the for-loop may be actually slower, but it's hard to guess. Can you please clarify how did you measure that performance improvement and how many methods the class you used for your benchmark has? Also it would be great to check what happens when you try to increase the number of those methods.