@vkrot 2016-02-26
I created a simple JMH test - https://github.com/vkrot/drools-pr-jmh
It fires rules for pojos with 10, 30 and 100 public methods, with and without property change listener.
I run it twice, with and without my patch
Results without patch:
```
DroolsBenchmark.testPojo100WithPropertyChangeListener     thrpt   60  110843,172 ± 4620,967  ops/s
DroolsBenchmark.testPojo100WithoutPropertyChangeListener  thrpt   60   90321,611 ± 3367,654  ops/s
DroolsBenchmark.testPojo10WitPropertyChangeListener       thrpt   60  111485,053 ± 4977,910  ops/s
DroolsBenchmark.testPojo10WithoutPropertyChangeListener   thrpt   60   91307,725 ± 3965,244  ops/s
DroolsBenchmark.testPojo30WithPropertyChangeListener      thrpt   60  112825,160 ± 5108,829  ops/s
DroolsBenchmark.testPojo30WithoutPropertyChangeListener   thrpt   60   92673,035 ± 3984,959  ops/s
```
results with patch:
```
DroolsBenchmark.testPojo100WithPropertyChangeListener     thrpt   60  127594,990 ± 3582,671  ops/s
DroolsBenchmark.testPojo100WithoutPropertyChangeListener  thrpt   60  125212,307 ± 3639,989  ops/s
DroolsBenchmark.testPojo10WitPropertyChangeListener       thrpt   60  189475,308 ± 8918,434  ops/s
DroolsBenchmark.testPojo10WithoutPropertyChangeListener   thrpt   60  183095,107 ± 9224,133  ops/s
DroolsBenchmark.testPojo30WithPropertyChangeListener      thrpt   60  171798,659 ± 6746,250  ops/s
DroolsBenchmark.testPojo30WithoutPropertyChangeListener   thrpt   60  172786,004 ± 7061,823  ops/s
```

to reproduce it check out benchmark and run 
```
mvn clean package
java -jar target/benchmarks.jar
```
As you can see, you are right, performance degrades as you add more methods. But even for really bad usecase (I suppose 100 methods per class is not very common) it still outperforms original version.
And it is still better for case when pojo does have property change listener, because `InitalFactImpl` is always inserted and it does not have property change listener.

By default this runs with patch. To run version without patch - delete ClassObjectTypeConf class.