@mrietveld 2016-02-26
`Class.getMethods()` is a relatively expensive operation since all superclasses have to be scanned going up the tree and put together in an array. I think that this case has not been taken into account in your performance tests? 

In cases in which the POJO has a deep hierarchy of parent tests, `Class.getMethods` will always be slower than `Class.getMethod(methodName, argTypes)`.