@mrietveld 2016-02-26
It may be worth looking at the following code: 

OpenJDK 6_14 [Class.getMethod(String name, Class<?>... parameterTypes)](http://grepcode.com/file/repository.grepcode.com/java/root/jdk/openjdk/6-b14/java/lang/Class.java#Class.getMethod%28java.lang.String%2Cjava.lang.Class[]%29)

OpenJDK 6_14 [Class.getMethods()](http://grepcode.com/file/repository.grepcode.com/java/root/jdk/openjdk/6-b14/java/lang/Class.java#Class.getMethods%28%29)