@mariofusco 2016-02-26
@mrietveld Sorry I just merged this PR before reading your comments :(
Do you think I should revert this?