@mariofusco 2016-02-26
Merged by https://github.com/droolsjbpm/drools/commit/d24cb6019