@mrietveld 2016-02-26
@mariofusco Personally, yes, I do not think this is a good idea and yes I think this should be reverted.

In short, this is code that tries to be smarter than the JDK itself. There's no way to tell if there are optimizations in the JDK (of which there are of course multiple implementations and versions) and which ones will be faster. Furthermore, (AFAIK?), this is not a pressing problem with regards to optimization. As Knuth has said: "Premature optimization is the root of all evil"...