@vkrot 2016-02-26
This code does not try to be smarter, I tried to eliminate throwing/catching exceptions at each call, because it is expensive to create exception, fill the stacktrace. I will find detailed benchmark later. Another solution - just cache already checked classes,  this will also eliminate exceptions and will not suffer from deep hierarchies problem. What do you think?